// frontend/src/api/generated/models/EmailRichHeader.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Todo Microservice API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface EmailRichHeader
 */
export interface EmailRichHeader {
  /**
   *
   * @type {string}
   * @memberof EmailRichHeader
   */
  from?: string;
  /**
   *
   * @type {string}
   * @memberof EmailRichHeader
   */
  subject?: string;
  /**
   *
   * @type {Date}
   * @memberof EmailRichHeader
   */
  date?: Date;
  /**
   *
   * @type {string}
   * @memberof EmailRichHeader
   */
  messageId?: string;
  /**
   *
   * @type {string}
   * @memberof EmailRichHeader
   */
  inReplyTo?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof EmailRichHeader
   */
  references?: Array<string>;
}

/**
 * Check if a given object implements the EmailRichHeader interface.
 */
export function instanceOfEmailRichHeader(value: object): value is EmailRichHeader {
  return true;
}

export function EmailRichHeaderFromJSON(json: any): EmailRichHeader {
  return EmailRichHeaderFromJSONTyped(json, false);
}

export function EmailRichHeaderFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): EmailRichHeader {
  if (json == null) {
    return json;
  }
  return {
    from: json['from'] == null ? undefined : json['from'],
    subject: json['subject'] == null ? undefined : json['subject'],
    date: json['date'] == null ? undefined : new Date(json['date']),
    messageId: json['messageId'] == null ? undefined : json['messageId'],
    inReplyTo: json['inReplyTo'] == null ? undefined : json['inReplyTo'],
    references: json['references'] == null ? undefined : json['references'],
  };
}

export function EmailRichHeaderToJSON(json: any): EmailRichHeader {
  return EmailRichHeaderToJSONTyped(json, false);
}

export function EmailRichHeaderToJSONTyped(
  value?: EmailRichHeader | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    from: value['from'],
    subject: value['subject'],
    date: value['date'] == null ? undefined : value['date'].toISOString(),
    messageId: value['messageId'],
    inReplyTo: value['inReplyTo'],
    references: value['references'],
  };
}